refactor: change naming for Prizepool component
